IO 메서드들에 In(Out)putStream/Reader/Writer 오버로딩 추가

+++ Pool.js
@@ -410,21 +410,38 @@ Pool.Canvas.drawCircle = function(center, radius, blockId, blockDamage){
 Pool.IO = {};
 
 /**
- * 문자열을 파일에 저장합니다
+ * 문자열을 파일이나 OutputStream 및 Writer에 저장합니다
  * 
  * @since 2015-02-26 (API 1)
  * @author affogatoman <colombia2@naver.com>
- * @param {String|File} path - 문자열을 저장할 파일의 경로 및 파일 객체
+ * @param {String|File|OutputStream|Writer} path - 문자열을 저장할 파일의 경로 및 파일 객체, OutputStream 및 Writer
+ * @param {String} str - 문자열을 저장할 파일의 경로 및 파일 객체
  */
-Pool.IO.saveFile = function(path, str){
+Pool.IO.saveAllLines = function(path, str){
 	try{
-		var file = path;
-		if(path instanceof String){
-			file = new java.io.File(path);
-		}
-		file.getParentFile().mkdirs();
+		var file = null;
+		var bw = null;
 		
-		var bw = new java.io.BufferedWriter(new java.io.FileWriter(file));
+		if(path instanceof java.io.File){
+			file = path;
+		}else if(path instanceof String){
+			file = new java.io.File(path);
+		}else if(path instanceof java.io.OutputStream){
+			file = null;
+			bw = new java.io.BufferedWriter(new java.io.FileWriter(new java.io.OutputStreamWriter(file)));
+		}else if(path instanceof java.io.Writer){
+			bw = path;
+		}else{
+			throw new Error("Illegal argument type");
+		}
+		
+		if(file !== null){
+			file.getParentFile().mkdirs();
+		}
+		if(bw === null){
+			bw = new java.io.BufferedWriter(new java.io.FileWriter(file));
+		}
+		
 		bw.write(str);
 		bw.close();
 	}catch(e){
@@ -433,33 +450,45 @@ Pool.IO.saveFile = function(path, str){
 };
 
 /**
- * 파일에서 문자열을 읽어옵니다
+ * 파일이나 InputStream 및 BufferedReader에서 문자열을 읽어옵니다
  * 
  * @since 2015-02-26 (API 1)
  * @author onebone <jyc0410@naver.com>
- * @param {String|File} path - 문자열을 읽을 파일의 경로 및 파일 객체
+ * @param {String|File|InputStream|BufferedReader} path - 문자열을 읽을 파일의 경로 및 파일 객체, InputStream 및 BufferedReader
  * @return {String} 파일 내의 문자열
  */
 Pool.IO.readAllLines = function(path){
 	try{
-		var file = path;
-		if(path instanceof String){
+		var file = null;
+		var br = null;
+		
+		if(path instanceof java.io.File){
+			file = path;
+		}else if(path instanceof String){
 			file = new java.io.File(path);
+		}else if(path instanceof java.io.InputStream){
+			file = null;
+			br = new java.io.BufferedReader(new java.io.FileReader(new java.io.InputStreamReader(file)));
+		}else if(path instanceof java.io.BufferedReader){
+			br = path;
+		}else{
+			throw new Error("Illegal argument type");
 		}
 		
-		if(!file.exists()){
+		if(file !== null && !file.exists()){
 			return null;
 		}
+		if(br === null){
+			new java.io.BufferedReader(new java.io.FileReader(file));
+		}
 		
-		var br = new java.io.BufferedReader(java.io.FileReader(file));
-		var sb = new java.lang.StringBuffer();
-		
-		var tmp;
+		var tmp = null, sb = new java.lang.StringBuffer();
 		while((tmp = br.readLine()) !== null){
 			sb.append(tmp);
 			sb.append("\n");
 		}
 		
+		br.close();
 		return sb.toString() + "";
 	}catch(e){
 		Pool.showError(e);
@@ -477,10 +506,15 @@ Pool.IO.readAllLines = function(path){
  */
 Pool.IO.removeFile = function(path){
 	try{
-		var file = path;
-		if(path instanceof String){
+		var file = null;
+		if(path instanceof java.io.File){
+			file = path;
+		}else if(path instanceof String){
 			file = new java.io.File(path);
+		}else{
+			throw new Error("Illegal argument type");
 		}
+		
 		return file["delete"](); //delete가 자바스크립트 키워드라...
 	}catch(e){
 		Pool.showError(e);
@@ -489,28 +523,51 @@ Pool.IO.removeFile = function(path){
 };
 
 /**
- * 주어진 파일을 다른 곳으로 복사합니다
+ * 주어진 파일이나 스트림을 다른 곳으로 복사합니다
  *
  * @since 2015-02-26 (API 1)
  * @author onebone <jyc0410@naver.com>
  * @author ChalkPE <amato0617@gmail.com>
- * @param {String|File} source - 원본 파일의 경로 혹은 파일 객체
+ * @param {String|File|InputStream} source - 원본 파일의 경로 혹은 파일 객체 및 InputStream
  * @param {String|File} target - 복사될 파일의 경로 혹은 파일 객체
  * @param {Boolean} deleteSourceAfterCopy - 복사 후 원본 파일을 삭제할 지의 여부
  * @return {Boolean} 파일 복시 성공 여부
  */
 Pool.IO.copyFile = function(source, target, deleteSourceAfterCopy){
-	if(source instanceof String){
-		source = new java.io.File(source);
-	}
-	if(target instanceof String){
-		target = new java.io.File(target);
-	}
-	
 	try{
+		var file = null;
+		var bis = null;
+		
+		if(source instanceof java.io.File){
+			file = source;
+		}else if(source instanceof String){
+			file = new java.io.File(source);
+		}else if(source instanceof java.io.InputStream){
+			bis = new java.io.BufferedInputStream(source);
+		}else{
+			throw new Error("Illegal argument type");
+		}
+		
+		if(file !== null && !file.exists()){
+			return false;
+		}
+		
+		if(target instanceof java.io.File){
+			target = target;
+		}else if(target instanceof String){
+			target = new java.io.File(target);
+		}else{
+			throw new Error("Illegal argument type");
+		}
+		
+		if(!target.exists()){
+			return false;
+		}
 		target.getParentFile().mkdirs();
 		
-		var bis = new java.io.BufferedInputStream(new java.io.FileInputStream(source));
+		if(bis === null){
+			bis = new java.io.BufferedInputStream(new java.io.FileInputStream(file));
+		}
 		var bos = new java.io.BufferedOutputStream(new java.io.FileOutputStream(target));
 		
 		var buf = java.lang.reflect.Array.newInstance(java.lang.Byte.TYPE, 4096);
